refactor: run VITS2 TensorRT in process
