+++ perception/config.yaml
@@ -33,8 +33,6 @@ plugins:
 
   vits2_tts_trt:
     enabled: true
-    backend_url: http://127.0.0.1:18080
-    request_timeout: 295
     speaker_id: 0
     speed: 1.0
     vits2_warmup: true
