+++ perception/plugins/vits2_tts_trt/requirements.jetson.txt
@@ -1,5 +1,3 @@
-fastapi==0.139.0
-uvicorn[standard]==0.51.0
 numpy==1.26.1
 pynini==2.1.6
 WeTextProcessing==1.0.4.1
